// src/app/@api/model/eventDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentDto } from './attachmentDto';
import { CityDto } from './cityDto';


export interface EventDto { 
    city: CityDto;
    content: string;
    eventDate: Date;
    eventEndDate: Date;
    eventId?: number;
    image?: AttachmentDto;
    status?: EventDto.StatusEnum;
    title: string;
}
export namespace EventDto {
    export type StatusEnum = 'PUBLISHED' | 'UNPUBLISHED';
    export const StatusEnum = {
        PUBLISHED: 'PUBLISHED' as StatusEnum,
        UNPUBLISHED: 'UNPUBLISHED' as StatusEnum
    };
}